Create new emoji-compat-flatbuffers module
This module only contains Apache 2 licensed flatbuffers code that
previously lived next to OFL and UNICODE licensed code in emoji-compat.

By extracting this, we can link the flatbuffers without having to add
notification of OFL and UNICODE in the module depending on these
flatbuffer definitions.

Update createfont.py to point to this new module, all files are
validated, and java src is moved to this new module.

Note: This change manually reverts the generated font since we are
missing source files for a reproducable build at time of merge. To be
addressed in followup CL.

Test: ./createfont.py {path}/NotoColorEmoji.ttf {path}/unicode ../emoji-compat-flatbuffers
Bug: b/178035684

Change-Id: I696debd751ad853c126a7b3d4d62db57db8cbc7d