Partial revert of 661d37fa4a85c038d069e5c11c0c24d2af77038a
This reverts the NotoSansLaoUI fonts which used to have LF glyph
but it was removed in 661d37fa4a85c038d069e5c11c0c24d2af77038a.
This causes a tofu rendering for LF character.
The right way to fix this is Roboto having a LF glyphs not to
fallback the other fonts, but updating Roboto is risky in
Android 13, thus reverting the LaoUI fonts for Android 13.

Bug: 235922054
Test: N/A
Change-Id: I14c57f622993203b15d07e3bc78c034fdd13c47c