Use the latest emoji for SDK for signing.
The droidcore target which includes the font files to the build
artifacts for signing but looks like it doesn't look the
soong_config_variables.
So, update the default font version to the latest one for
proceeding the font signing process.

This doesn't change the font file actually installed in the
system image.

Bug: 358182138
Test: N/A
Flag: None (build only change)
Change-Id: Icab9ac1b794550761a34ccde6fdeff14002a4d64